Add a tagged HashEngine
Currently the `sha256t` module uses an alias to the
`sha256::HashEngine`, this is a footgun because if a user accidentally
uses the `sha256t` module the same way as the other hash modules they
will get the wrong engine. For example this code looks fine but is
buggy (it does not pre-tag the engine):

```
 let mut engine = sha256t::HashEngine::new();  // Or `default()`.
 engine.input(some_data);
 let hash = engine.finalize();
```

We can fix the situation by adding a `HashEngine<T>` type to the
`sha256t` module and implementing `Default` for it by calling the
`Tag::engine` function.