Rename hmac types
The `hmac` types are named differently to the other hash types. At this
stage there is not that much difference between them and it is not
immediately obvious why they get a different naming scheme.

As we do for the other hash types use `Hash` and `HashEngine` with the
expectation that user differentiate by using the module path.

With this patch applied users use `hmac::Hash` instead of `Hmac`.